Extract common optimization rules for QL
